+++ finagle-http/src/test/scala/com/twitter/finagle/http/HttpMuxerTest.scala
@@ -1,16 +1,15 @@
 package com.twitter.finagle.http
 
+import com.twitter.conversions.DurationOps._
 import com.twitter.finagle.http.Method.Patch
 import com.twitter.finagle.Service
 import com.twitter.io.Buf
-import com.twitter.util.{Await, Future}
-import org.junit.runner.RunWith
+import com.twitter.util.{Await, Future, Time}
 import org.scalatest.FunSuite
-import org.scalatest.junit.JUnitRunner
 
-@RunWith(classOf[JUnitRunner])
 class HttpMuxerTest extends FunSuite {
-  // todo: add other metrics when they are supported
+  def await[A](f: Future[A]): A = Await.result(f, 5.seconds)
+
   class DummyService(reply: String) extends Service[Request, Response] {
     def apply(request: Request): Future[Response] = {
       val response = Response(Version.Http11, Status.Ok)
@@ -34,45 +33,45 @@ class HttpMuxerTest extends FunSuite {
   }
 
   test("prefix matching is handled correctly") {
-    assert(Await.result(muxService(Request("/foo/<a>"))).contentString == percentEncode)
+    assert(await(muxService(Request("/foo/<a>"))).contentString == percentEncode)
 
-    assert(Await.result(muxService(Request("/fooblah"))).status == Status.NotFound)
+    assert(await(muxService(Request("/fooblah"))).status == Status.NotFound)
 
-    assert(Await.result(muxService(Request("/foo/bar/blah"))).contentString == fooBarPrefix)
+    assert(await(muxService(Request("/foo/bar/blah"))).contentString == fooBarPrefix)
 
-    assert(Await.result(muxService(Request("/foo//bar/blah"))).contentString == fooBarPrefix)
+    assert(await(muxService(Request("/foo//bar/blah"))).contentString == fooBarPrefix)
 
-    assert(Await.result(muxService(Request("/foo//bar/<a>"))).contentString == fooBarPrefix)
+    assert(await(muxService(Request("/foo//bar/<a>"))).contentString == fooBarPrefix)
 
-    assert(Await.result(muxService(Request("/foo/bar"))).contentString == fooBarExact)
+    assert(await(muxService(Request("/foo/bar"))).contentString == fooBarExact)
 
-    assert(Await.result(muxService(Request("/foo/bar/"))).contentString == fooBarPrefix)
+    assert(await(muxService(Request("/foo/bar/"))).contentString == fooBarPrefix)
 
-    assert(Await.result(muxService(Request("/foo/boo/baz"))).status == Status.NotFound)
+    assert(await(muxService(Request("/foo/boo/baz"))).status == Status.NotFound)
 
-    assert(Await.result(muxService(Request("/foo/boo/baz/blah"))).contentString == fooBooBaz)
+    assert(await(muxService(Request("/foo/boo/baz/blah"))).contentString == fooBooBaz)
 
-    assert(Await.result(muxService(Request("/foo/barblah"))).status == Status.NotFound)
+    assert(await(muxService(Request("/foo/barblah"))).status == Status.NotFound)
   }
 
   test("exact matching is handled correctly") {
-    assert(Await.result(muxService(Request("/exact/match"))).contentString == exactMatch)
+    assert(await(muxService(Request("/exact/match"))).contentString == exactMatch)
 
-    assert(Await.result(muxService(Request("/exact/match/"))).status == Status.NotFound)
+    assert(await(muxService(Request("/exact/match/"))).status == Status.NotFound)
 
-    assert(Await.result(muxService(Request("/exact/match/nested"))).status == Status.NotFound)
+    assert(await(muxService(Request("/exact/match/nested"))).status == Status.NotFound)
   }
 
   test("""special cases "" and "/" are handled correctly""") {
     val slashMux = new HttpMuxer().withHandler("/", new DummyService(specialCase))
-    assert(Await.result(slashMux(Request("/"))).contentString == specialCase)
-    assert(Await.result(slashMux(Request(""))).status == Status.NotFound)
-    assert(Await.result(slashMux(Request("/anything"))).contentString == specialCase)
+    assert(await(slashMux(Request("/"))).contentString == specialCase)
+    assert(await(slashMux(Request(""))).status == Status.NotFound)
+    assert(await(slashMux(Request("/anything"))).contentString == specialCase)
 
     val emptyStringMux = new HttpMuxer().withHandler("", new DummyService(specialCase))
-    assert(Await.result(emptyStringMux(Request("/"))).contentString == specialCase)
-    assert(Await.result(emptyStringMux(Request(""))).contentString == specialCase)
-    assert(Await.result(emptyStringMux(Request("/anything"))).status == Status.NotFound)
+    assert(await(emptyStringMux(Request("/"))).contentString == specialCase)
+    assert(await(emptyStringMux(Request(""))).contentString == specialCase)
+    assert(await(emptyStringMux(Request("/anything"))).status == Status.NotFound)
   }
 
   test("Registering a service with an existing name will overwrite the old") {
@@ -80,8 +79,8 @@ class HttpMuxerTest extends FunSuite {
     val orig = new HttpMuxer().withHandler("foo/", new DummyService(origResp))
     val overridden = orig.withHandler("foo/", new DummyService(newResp))
 
-    assert(Await.result(orig(Request("/foo/bar"))).contentString == origResp)
-    assert(Await.result(overridden(Request("/foo/bar"))).contentString == newResp)
+    assert(await(orig(Request("/foo/bar"))).contentString == origResp)
+    assert(await(overridden(Request("/foo/bar"))).contentString == newResp)
   }
 
   test("RouteIndex only allows GET and POST") {
@@ -92,5 +91,19 @@ class HttpMuxerTest extends FunSuite {
         method = Patch
       )
     }
+  }
+
+  test("Closing the HttpMuxer closes the individual handlers") {
+    var closed = false
+    val closableSvc = new Service[Request, Response] {
+      def apply(req: Request): Future[Response] = ???
+      override def close(deadline: Time): Future[Unit] = {
+        closed = true
+        Future.Done
+      }
+    }
+    val closable = new HttpMuxer().withHandler("foo/", closableSvc)
+    await(closable.close())
+    assert(closed)
   }
 }